feat: 폴더구조 조정 (#163)
